Crate autocxx_macro

Macros

include_cpp_impl

Implementation of the include_cpp macro. See documentation for autocxx crate.

Attribute Macros

cpp_semantics

Attribute which should never be encountered in real life. This is something which features in the Rust source code generated by autocxx-bindgen and passed to autocxx-engine, which should never normally be compiled by rustc before it undergoes further processing.

extern_rust_function

Attribute to state that a Rust function is to be exported to C++ in the extern "Rust" section of the generated cxx bindings.

extern_rust_type

Attribute to state that a Rust type is to be exported to C++ in the extern "Rust" section of the generated cxx bindings.

subclass

Attribute to state that a Rust struct is a C++ subclass. This adds an additional field to the struct which autocxx uses to track a C++ instantiation of this Rust subclass.